Added comment and event flagging
Added support for comment and event flagging

Simply post to either /api/reports/comments or /api/reports/events
with the content id and the reason for the flagging.
You have to be logged in to be able to flag (so you need an access token).

TODO: Validate the input map a bit more, add more tests